fix: bugfix for router dialog
