+++ packages/plugins/SmartPay/src/locales/en-US.json
@@ -37,7 +37,8 @@
     "ineligibility_tips": "Sorry, You are not in the free trial whitelist.",
     "personas_description": "Your Twitter can create one Smart Pay wallet.",
     "swap": "Swap",
-    "smart_pay_wallet": "Smart Pay",
+    "smart_pay_wallet": "Smart Pay Wallet",
+    "smart_pay_wallet_deployment": "Smart Pay wallet deployment",
     "setup_smart_pay_managed_account": "Set up Smart Pay's Managed Account",
     "send": "Send",
     "smart_pay_description": "Smart Pay is based on ERC 4337 and is currently deployed on Polygon network. Multiple chains will be supported in the future. ERC 4337 is a new wallet proposal that would open the door for creativity in wallet design and could provide more important features. And we will design more fascinating features based on the new wallet proposal in Mask Network.",
